+++ trane/utils/__init__.py
@@ -1,4 +1,3 @@
-from trane.utils.data_parser import *  # noqa
 from trane.utils.featuretools_wrapper import *  # noqa
 from trane.utils.helper import (
     _solve_single_problem,